Update AnuraUserModel to change weight type from double to int and modify AnuraScannerActivity to handle weight as an integer. Ensure user data extraction is safe with non-null assertions.

+++ anura_flutter/lib/models/anura_user_model.dart
@@ -18,7 +18,7 @@ class AnuraUserModel {
   final int age;
 
   /// Weight in KG
-  final double weight;
+  final int weight;
 
   /// To create json from [AnuraUserModel]
   Map<String, dynamic> toJson() => {
